backend/requirements.txt
asgiref==3.11.0
Django==5.1.4
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
sqlparse==0.5.5
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0
markdown==3.7
pytz==2024.2
